migrate @klimadao/site to import from lib
