+++ lib/components/index.ts
@@ -25,4 +25,4 @@ export { LogoWithClaim } from "./Logos/LogoWithClaim";
 export { Text } from "./Text";
 export { ThemeToggle } from "./ThemeToggle";
 export { InitializeTheme } from "./InitializeTheme";
-export { Web3ContextProvider, useWeb3 } from "./Web3Context";
+export { Web3ContextProvider } from "./Web3Context";
